move query builder to own file
