+++ docs/Toy/toy/rewrites/optimise_toy.py
@@ -13,9 +13,9 @@ from xdsl.pattern_rewriter import (
     PatternRewriter,
     PatternRewriteWalker,
 )
+from xdsl.rewriting.query_builder import rewrite_pattern_query
 from xdsl.rewriting.sasha_rewrite_pattern import (
     query_rewrite_pattern,
-    rewrite_pattern_query,
 )
 from xdsl.transforms.dead_code_elimination import dce
 from xdsl.utils.hints import isa
